Move all the work to build aggregators to `_agg.py`
